Add the create user endpoint
