+++ app/api/cruds/login.py
@@ -1,21 +1,8 @@
 from sqlalchemy.orm import Session
 
+from app.api.cruds.user import get_user_by_email
 from app.api.models.user import User
 from app.api.utils.security import verify_password
-
-
-def get_user_by_email(db: Session, email: str) -> User | None:
-    """
-    Retrieves a user from the database based on their email.
-
-    Args:
-        db: The database session.
-        email: The email of the user.
-
-    Returns:
-        The user object if found, otherwise None.
-    """
-    return db.query(User).filter(User.email == email).first()
 
 
 def authenticate(db: Session, email: str, password: str) -> User | None:
