+++ app/api/cruds/user.py
@@ -55,3 +55,17 @@ def get_user_by_id(db: Session, id: int) -> User | None:
         The user object if found, otherwise None.
     """
     return db.query(User).filter(User.id == id).first()
+
+
+def get_user_by_email(db: Session, email: str) -> User | None:
+    """
+    Retrieves a user from the database based on their email.
+
+    Args:
+        db: The database session.
+        email: The email of the user.
+
+    Returns:
+        The user object if found, otherwise None.
+    """
+    return db.query(User).filter(User.email == email).first()
